Completed auth and exceptions

+++ backend/database-service/src/main/java/com/elms/databaseservice/controllers/FeedbackController.java
@@ -1,9 +1,11 @@
 package com.elms.databaseservice.controllers;
 
+import java.util.Collections;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -36,40 +38,56 @@ public class FeedbackController {
 
 
 	@GetMapping(path = "/feedbacks")
-	public List<Feedback> fetchAllStudents(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
+	public ResponseEntity<List<Feedback>> fetchAllStudents(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader) throws Exception {
 		if(client.authorizeTheRequest(requestTokenHeader))
-		return service.getAllFeedbacks();
+			return service.getAllFeedbacks();
 		else
-			throw new Exception("No feedbacks avialable");
+			return new ResponseEntity<List<Feedback>>(Collections.EMPTY_LIST,HttpStatus.BAD_REQUEST);
 
 	}
 
 	@GetMapping(path = "/course/{courseId}/feedback")
-	public ResponseEntity<List<Feedback>> fetchAllFeedbackByCourseId(@PathVariable("courseId") int courseId) {
-		return service.getAllFeedbacksByCourseId(courseId);
+	public ResponseEntity<List<Feedback>> fetchAllFeedbackByCourseId(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader,@PathVariable("courseId") int courseId) {
+		if(client.authorizeTheRequest(requestTokenHeader))
+			return service.getAllFeedbacksByCourseId(courseId);
+		else
+			return new ResponseEntity<List<Feedback>>(Collections.EMPTY_LIST,HttpStatus.BAD_REQUEST);
 	}
 
 	@GetMapping(path = "/student/{studentId}/course/{courseId}/feedback")
-	public ResponseEntity<String> isFeedbackPresent(@PathVariable("studentId") int studentId,
+	public ResponseEntity<String> isFeedbackPresent(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader,@PathVariable("studentId") int studentId,
 			@PathVariable("courseId") int courseId) {
-		return service.existFeedbackById(studentId, courseId);
+		if(client.authorizeTheRequest(requestTokenHeader))
+			return service.existFeedbackById(studentId, courseId);
+		else
+			return new ResponseEntity<>("User not authenticated",HttpStatus.BAD_REQUEST);
 	}
 
 //
 	@PostMapping(path = "/student/{id}/course/{courseId}/feedback")
-	public ResponseEntity<String> storeFeedback(@PathVariable("id") int id,
+	public ResponseEntity<String> storeFeedback(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader,@PathVariable("id") int id,
 			@PathVariable("courseId") int courseId,
 			@RequestBody Feedback feedback) {
-		StudentCourse studentCourse=studentCourseRepo.findById(new StudentCourseId(id,courseId)).get();
-		feedback.setStudentCourseId(studentCourse);
-		 return service.storeFeedback(feedback);
+		if(client.authorizeTheRequest(requestTokenHeader))
+		{
+			StudentCourse studentCourse=studentCourseRepo.findById(new StudentCourseId(id,courseId)).get();
+			feedback.setStudentCourseId(studentCourse);
+			return service.storeFeedback(feedback);
+		}
+		else
+			return new ResponseEntity<>("User not authenticated",HttpStatus.BAD_REQUEST);
+		
 	}
 
 	@DeleteMapping(path = "/student/{studentId}/course/{courseId}/feedback")
-	public ResponseEntity<String> deleteFeedback(@PathVariable("studentId") int studentId,
+	public ResponseEntity<String> deleteFeedback(@RequestHeader(value = "Authorization", required = true) String requestTokenHeader,@PathVariable("studentId") int studentId,
 			@PathVariable("courseId") int courseId) {
 		logger.info("feedback savings");
-		return service.deleteFeedback(studentId, courseId);
+		if(client.authorizeTheRequest(requestTokenHeader))
+			return service.deleteFeedback(studentId, courseId);
+		else
+			return new ResponseEntity<>("User not authenticated",HttpStatus.BAD_REQUEST);
+	
 	}
 
 }
